requirements.txt
requests
python-dotenv
pydantic
rootpath

streamlit
openai
